run generate with new features
