+++ prisma/schema.prisma
@@ -4,6 +4,7 @@ generator client {
 
 generator docs {
   provider = "node ./dist/index.js"
+  typeSuffix = ""
 }
 
 datasource db {
