+++ prisma/typebox/Post.ts
@@ -1,8 +1,11 @@
 import { Type, Static } from "@sinclair/typebox";
 
-export const Post = Type.Object({
-  id: Type.Number(),
-  userId: Type.Optional(Type.Number()),
-});
+export const Post = Type.Object(
+  {
+    id: Type.Number(),
+    userId: Type.Optional(Type.Number()),
+  },
+  { $id: "Post" },
+);
 
-export type PostType = Static<typeof Post>;
+export type Post = Static<typeof Post>;
