+++ prisma/typebox/PostInput.ts
@@ -1,8 +1,11 @@
 import { Type, Static } from "@sinclair/typebox";
 
-export const PostInput = Type.Object({
-  id: Type.Optional(Type.Number()),
-  userId: Type.Optional(Type.Number()),
-});
+export const PostInput = Type.Object(
+  {
+    id: Type.Optional(Type.Number()),
+    userId: Type.Optional(Type.Number()),
+  },
+  { $id: "PostInput" },
+);
 
-export type PostInputType = Static<typeof PostInput>;
+export type PostInput = Static<typeof PostInput>;
